teach `debug.ElfSymTab` to read unwind info in `.eh_frame` sections
This makes it possible for executables built with `-fstrip=debuginfo` to unwind
frames. No need to specify `-fno-omit-frame-pointer`! This is only possible
because the `eh_frame` and `eh_frame_hdr` sections are not stripped when
`-fstrip=debuginfo`.

Also, remove check in MachO code for `error.RequiresDWARFUnwind`. That
error was removed[1] way back in 2023, but wasn't cleaned up properly.

[1]: 97bda56306622784b2f2e4f036d817bbe5f2025e