cmake, bootstrap.c: Set a 16 MiB stack size for zig1.
This is the default stack size that the Zig compiler sets when linking Zig
programs. Since we're compiling a Zig program in C form here, it seems sensible
to do the same.

This might need to be increased even more for compilers like tcc and slimcc, as
is done for zig2, but that can come later.