mark std.testing.print as a public function
Zig's stdlib testing is able to print helpful error messages when
attempting to validate a test. This works both during comptime, and
runtime. This exposes the previously internal function to be used in any
custom built test functions to produce helpful output that is similar
behavior to the standard test framework.