std.Target: Mark Cpu.Arch.endian() inline.
Before this change, an explicit comptime annotation was required to turn a call
to builtin.target.cpu.arch.endian() into a guaranteed constant. A quick grep of
the codebase shows that this annotation was forgotten in quite a few places,
resulting in a pointless runtime call to endian() along with potentially dead
code ending up in the binary, depending on the LLVM inliner's mood.

Just mark this function inline since, like many other similar functions in
std.Target, people clearly expect it to be.

This results in a larger Debug compiler because of inlined calls on non-constant
std.Target.Cpu.Arch values. However, a ReleaseFast compiler is smaller, which is
presumably because LLVM deduplicates the switch tables for non-constant endian()
calls, and is able to optimize based on the value of constant endian() calls.